backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.1
stripe>=2.61.0
supabase>=0.5.8
pillow>=8.4.0
requests>=2.26.0 
